+++ src/exception/Message.php
@@ -68,6 +68,14 @@ enum Message: string
     case VALUE_NOT_IN_LIST = "Value '%s' is not in the list of valid values for '%s': '%s'.";
 
     /**
+     * Error when a value is of an invalid type.
+     *
+     * Format: "Value should be of type 'array', 'scalar', 'null', or 'enum'; '%s' given."
+     */
+    case VALUE_SHOULD_BE_ARRAY_SCALAR_NULL_ENUM = "Value should be of type 'array', 'scalar', 'null', or 'enum'; " .
+    "'%s' given.";
+
+    /**
      * Returns the formatted message string for the error case.
      *
      * Retrieves and formats the error message string by interpolating the provided arguments.